requirements.txt
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-xdist==3.7.0
mcp_use==1.3.1
langchain-openai==0.3.23
python-dotenv==1.1.0
fastmcp==2.6.1
